+++ CHANGELOG.md
@@ -83,6 +83,8 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
   * OS distribution support minimum for roofline feature is now Ubuntu22.04, RHEL9, and SLES15SP6
 * Improve analysis block based filtering to accept metric id level filtering
   * This can be used to collect individual metrics from various sections of analysis config
+* CLI analysis mode baseline comparison will now only compare common metrics across workloads and will not show Metric ID
+  * Remove metrics from analysis configuration files which are explicitly marked as empty or None
 
 ### Optimized